% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/witcher_spells.R
\docType{data}
\name{witcher.map}
\alias{witcher.map}
\title{'The Witcher' colour map}
\format{
A data frame containing all the colours used in the palette:
\itemize{
\item V1: Red value
\item V2: Green value
\item V3: Blue value
\item option: Refers to the characters of the world of the Witcher.
}
}
\usage{
witcher.map
}
\description{
A dataset containing the colour palettes from the TV show, book series and video games of 'The Witcher'.
}
\keyword{datasets}
